# docs/evidencias/openapi-v3-api-docs.json
{"openapi":"3.1.0","info":{"title":"ServiceHub API","description":"API RESTful do ServiceHub, plataforma de catálogo de serviços. Nesta etapa (Checkpoint), o recurso **Serviços** possui CRUD completo. Os erros seguem o padrão Problem Details (RFC 9457).","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Serviços","description":"Operações de cadastro, consulta, atualização e exclusão de serviços oferecidos na plataforma ServiceHub."}],"paths":{"/api/v1/servicos":{"get":{"tags":["Serviços"],"summary":"Listar todos os serviços","description":"Retorna a lista com todos os serviços cadastrados, ordenados pelo identificador. Quando não há serviços cadastrados, retorna uma lista vazia com status 200.","operationId":"listar","responses":{"200":{"description":"Lista de serviços (pode estar vazia)","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServicoResponse"}}}}}}},"post":{"tags":["Serviços"],"summary":"Cadastrar um novo serviço","description":"Cria um serviço a partir dos dados informados no corpo da requisição. Nome, categoria, preço e duração estimada são obrigatórios; a descrição é opcional. O identificador e as datas de criação e atualização são gerados pela API. Em caso de sucesso, o cabeçalho Location contém a URI do novo recurso.","operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicoRequest"}}},"required":true},"responses":{"201":{"description":"Serviço criado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicoResponse"}}}},"400":{"description":"Dados inválidos: campo obrigatório ausente, valor fora dos limites ou JSON malformado","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"examples":{"Dados inválidos":{"description":"Dados inválidos","value":{"type":"about:blank","title":"Dados inválidos","status":400,"detail":"Um ou mais campos são inválidos","instance":"/api/v1/servicos","erros":[{"campo":"preco","mensagem":"O preço deve ser maior que zero"}]}}}}}}}}},"/api/v1/servicos/{id}":{"get":{"tags":["Serviços"],"summary":"Buscar um serviço por ID","description":"Retorna os dados completos do serviço identificado pelo ID informado na URI.","operationId":"buscarPorId","parameters":[{"name":"id","in":"path","description":"Identificador do serviço","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Serviço encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicoResponse"}}}},"400":{"description":"ID em formato inválido (não numérico)","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Nenhum serviço encontrado com o ID informado","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"examples":{"Serviço inexistente":{"description":"Serviço inexistente","value":{"type":"about:blank","title":"Recurso não encontrado","status":404,"detail":"Serviço com id 99 não encontrado","instance":"/api/v1/servicos/99"}}}}}}}},"put":{"tags":["Serviços"],"summary":"Atualizar um serviço","description":"Substitui todos os dados do serviço identificado pelo ID (semântica do PUT): o corpo deve conter a representação completa, com os mesmos campos obrigatórios do cadastro. Se a descrição for omitida, ela é limpa. A data de criação é preservada e a data de atualização é renovada.","operationId":"atualizar","parameters":[{"name":"id","in":"path","description":"Identificador do serviço a ser atualizado","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicoRequest"}}},"required":true},"responses":{"200":{"description":"Serviço atualizado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicoResponse"}}}},"400":{"description":"Dados inválidos ou ID em formato inválido","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Nenhum serviço encontrado com o ID informado","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}},"delete":{"tags":["Serviços"],"summary":"Excluir um serviço","description":"Remove permanentemente o serviço identificado pelo ID. Em caso de sucesso, a resposta não possui corpo (204 No Content).","operationId":"excluir","parameters":[{"name":"id","in":"path","description":"Identificador do serviço a ser excluído","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"204":{"description":"Serviço excluído com sucesso (sem corpo na resposta)"},"400":{"description":"ID em formato inválido (não numérico)","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Nenhum serviço encontrado com o ID informado","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}}},"components":{"schemas":{"ServicoRequest":{"type":"object","description":"Dados para criação ou atualização completa de um serviço","properties":{"nome":{"type":"string","description":"Nome do serviço","example":"Instalação de ar-condicionado","maxLength":100,"minLength":3},"descricao":{"type":["string","null"],"description":"Descrição detalhada do serviço (opcional)","example":"Instalação de split de até 12.000 BTUs, com suporte e tubulação de até 3 metros","maxLength":500,"minLength":0},"categoria":{"type":"string","description":"Categoria em que o serviço se enquadra","example":"Climatização","maxLength":60,"minLength":0},"preco":{"type":"number","description":"Preço do serviço em reais (R$), com até 2 casas decimais","example":350.0,"maximum":99999999.99,"minimum":0.01},"duracaoMinutos":{"type":"integer","format":"int32","description":"Duração estimada do serviço, em minutos","example":120,"maximum":100000,"minimum":1}},"required":["categoria","duracaoMinutos","nome","preco"]},"ServicoResponse":{"type":"object","description":"Representação de um serviço cadastrado","properties":{"id":{"type":"integer","format":"int64","description":"Identificador único do serviço","example":1},"nome":{"type":"string","description":"Nome do serviço","example":"Instalação de ar-condicionado"},"descricao":{"type":["string","null"],"description":"Descrição detalhada do serviço","example":"Instalação de split de até 12.000 BTUs, com suporte e tubulação de até 3 metros"},"categoria":{"type":"string","description":"Categoria do serviço","example":"Climatização"},"preco":{"type":"number","description":"Preço do serviço em reais (R$)","example":350.0},"duracaoMinutos":{"type":"integer","format":"int32","description":"Duração estimada, em minutos","example":120},"criadoEm":{"type":"string","format":"date-time","description":"Data e hora de criação (UTC, ISO-8601)","example":"2026-09-20T15:04:05.123Z"},"atualizadoEm":{"type":"string","format":"date-time","description":"Data e hora da última atualização (UTC, ISO-8601)","example":"2026-09-20T15:04:05.123Z"}}},"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":["string","null"]},"status":{"type":"integer","format":"int32"},"detail":{"type":["string","null"]},"instance":{"type":["string","null"],"format":"uri"},"properties":{"type":["object","null"],"additionalProperties":{}}}}}}}